+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -1,23 +1,38 @@
 package com.techcourse.controller;
 
 import com.techcourse.domain.User;
+import com.techcourse.domain.UserSession;
 import com.techcourse.repository.InMemoryUserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-import nextstep.mvc.controller.asis.Controller;
 import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 import nextstep.mvc.view.View;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.support.RequestMethod;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class LoginController implements Controller {
+@Controller
+public class LoginController {
 
-    private static final Logger log = LoggerFactory.getLogger(LoginController.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(LoginController.class);
 
-    @Override
-    public ModelAndView execute(HttpServletRequest req, HttpServletResponse res) {
+    @RequestMapping(value = "/login/view", method = RequestMethod.GET)
+    public ModelAndView view(HttpServletRequest req, HttpServletResponse res) {
+        return UserSession.getUserFrom(req.getSession())
+            .map(user -> {
+                LOGGER.info("logged in {}", user.getAccount());
+                View view = new JspView("redirect:/index.jsp");
+                return new ModelAndView(view);
+            })
+            .orElse(new ModelAndView(new JspView("/login.jsp")));
+    }
+
+    @RequestMapping(value = "/login", method = RequestMethod.POST)
+    public ModelAndView login(HttpServletRequest req, HttpServletResponse res) {
         if (UserSession.isLoggedIn(req.getSession())) {
             View view = new JspView("redirect:/index.jsp");
             return new ModelAndView(view);
@@ -25,15 +40,15 @@ public class LoginController implements Controller {
 
         return InMemoryUserRepository.findByAccount(req.getParameter("account"))
             .map(user -> {
-                log.info("User : {}", user);
-                return new ModelAndView(login(req, user));
+                LOGGER.info("User : {}", user);
+                return new ModelAndView(loginCheck(req, user));
             })
             .orElse(new ModelAndView(
                 new JspView("redirect:/401.jsp")
             ));
     }
 
-    private View login(HttpServletRequest request, User user) {
+    private View loginCheck(HttpServletRequest request, User user) {
         if (user.checkPassword(request.getParameter("password"))) {
             final HttpSession session = request.getSession();
             session.setAttribute(UserSession.SESSION_KEY, user);
