--- app/src/main/java/com/techcourse/controller/LoginViewController.java
@@ -1,26 +1,0 @@
-package com.techcourse.controller;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.view.JspView;
-import nextstep.mvc.view.ModelAndView;
-import nextstep.mvc.view.View;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class LoginViewController implements Controller {
-
-    private static final Logger log = LoggerFactory.getLogger(LoginViewController.class);
-
-    @Override
-    public ModelAndView execute(HttpServletRequest req, HttpServletResponse res) {
-        return UserSession.getUserFrom(req.getSession())
-                .map(user -> {
-                    log.info("logged in {}", user.getAccount());
-                    View view = new JspView("redirect:/index.jsp");
-                    return new ModelAndView(view);
-                })
-                .orElse(new ModelAndView(new JspView("/login.jsp")));
-    }
-}
